+++ Marlin/src/module/scara.cpp
@@ -49,10 +49,10 @@ float segments_per_second = DEFAULT_SEGMENTS_PER_SECOND;
    * Integrated into Marlin and slightly restructured by Joachim Cerny.
    */
   void forward_kinematics(const_float_t a, const_float_t b) {
-    const float a_sin = sin(RADIANS(a)) * L1,
-                a_cos = cos(RADIANS(a)) * L1,
-                b_sin = sin(RADIANS(SUM_TERN(MP_SCARA, b, a))) * L2,
-                b_cos = cos(RADIANS(SUM_TERN(MP_SCARA, b, a))) * L2;
+    const float a_sin = sinf(RADIANS(a)) * L1,
+                a_cos = cosf(RADIANS(a)) * L1,
+                b_sin = sinf(RADIANS(SUM_TERN(MP_SCARA, b, a))) * L2,
+                b_cos = cosf(RADIANS(SUM_TERN(MP_SCARA, b, a))) * L2;
 
     cartes.x = a_cos + b_cos + scara_offset.x;  // theta
     cartes.y = a_sin + b_sin + scara_offset.y;  // phi
@@ -199,10 +199,10 @@ float segments_per_second = DEFAULT_SEGMENTS_PER_SECOND;
   // Convert ABC inputs in degrees to XYZ outputs in mm
   void forward_kinematics(const_float_t a, const_float_t b, const_float_t c) {
     const float w = c - b,
-                r = L1 * cos(RADIANS(b)) + L2 * sin(RADIANS(w - (90 - b))),
-                x = r  * cos(RADIANS(a)),
-                y = r  * sin(RADIANS(a)),
-                rho2 = L1_2 + L2_2 - 2.0f * L1 * L2 * cos(RADIANS(w));
+                r = L1 * cosf(RADIANS(b)) + L2 * sinf(RADIANS(w - (90 - b))),
+                x = r  * cosf(RADIANS(a)),
+                y = r  * sinf(RADIANS(a)),
+                rho2 = L1_2 + L2_2 - 2.0f * L1 * L2 * cosf(RADIANS(w));
 
     cartes = robot_offset + xyz_pos_t({ x, y, SQRT(rho2 - sq(x) - sq(y)) });
   }
@@ -285,7 +285,7 @@ float segments_per_second = DEFAULT_SEGMENTS_PER_SECOND;
                 GAMMA = ATAN2(SG, CG), // Method 2
 
                 // Angle of Shoulder Joint, elevation angle measured from horizontal (r+)
-                //PHI = asin(spos.z/RHO) + asin(L2 * sin(GAMMA) / RHO),
+                //PHI = asinf(spos.z/RHO) + asinf(L2 * sinf(GAMMA) / RHO),
                 PHI = ATAN2(spos.z, RXY) + ATAN2(K2, K1),   // Method 2
 
                 // Elbow motor angle measured from horizontal, same frame as shoulder  (r+)
